% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-markdown.R
\name{ragnar_chunks_view}
\alias{ragnar_chunks_view}
\title{View chunks with the store inspector}
\usage{
ragnar_chunks_view(chunks)
}
\arguments{
\item{chunks}{A data frame containing a few chunks.}
}
\description{
Visualize chunks read by \code{\link[=ragnar_read]{ragnar_read()}} for quick inspection.
Helpful for inspecting the results of chunking and reading while iterating
on the ingestion pipeline.
}
