% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{ragnar_embed_ollama}
\alias{ragnar_embed_ollama}
\title{Embedd Text}
\usage{
ragnar_embed_ollama(
  x,
  base_url = "http://localhost:11434",
  model = "all-minilm",
  batch_size = 10L
)
}
\arguments{
\item{x}{A character vector, or a data frame with a column named \code{"text"}}

\item{base_url}{url where the ollama service is available}

\item{model}{string; model name}

\item{batch_size}{split \code{x} into batches when embedding.}
}
\value{
If \code{x} is a character vector, then a numeric matrix is returned,
where \code{nrow = length(x)} and \verb{ncol = <model-embedding-size>}.
If \code{x} is a data.frame, then a new \code{embedding} matrix "column" is added,
containing the matrix described in the previous sentence.
}
\description{
Embedd Text
}
\examples{
text <- c("a chunk of text", "another chunk of text", "one more chunk of text")
text |>
  ragnar_embed_ollama() |>
  str()
}
