% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{ragnar_embed_openai}
\alias{ragnar_embed_openai}
\title{Embed text via OpenAI API}
\usage{
ragnar_embed_openai(
  x,
  model = "text-embedding-3-small",
  base_url = "https://api.openai.com/v1",
  api_key = openai_key(),
  dims = NULL,
  user = get_ragnar_username(),
  batch_size = 20
)
}
\arguments{
\item{x}{character vector, or a dataframe with a column named \code{text}.}

\item{model}{string}

\item{base_url}{api url}

\item{api_key}{resolved using env var \code{OPENAI_API_KEY}}

\item{dims}{An integer, can be used to truncate the embedding to a specific size.}

\item{user}{User name passed via the API.}

\item{batch_size}{Integer, limit of strings to include in a single request.}
}
\value{
A matrix of embeddings with 1 row per input string, or a dataframe with an 'embedding' column.
}
\description{
Embed text via OpenAI API
}
