% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-markdown.R
\name{ragnar_read}
\alias{ragnar_read}
\title{Read a document as Markdown}
\usage{
ragnar_read(paths, ...)
}
\arguments{
\item{...}{passed on \code{markitdown.convert}.}

\item{x}{file path or url.}
}
\value{
Always returns a data frame with the columns:
\itemize{
\item \code{origin}: the file path or url
\item \code{text}: the markdown content
}
}
\description{
\code{ragnar_read()} uses \href{https://github.com/microsoft/markitdown}{markitdown} to
convert a document to markdown.
}
