% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ragnar_store_update_document_and_chunks}
\alias{ragnar_store_update_document_and_chunks}
\title{Workhorse that inserts a document and its chunks into the store.
Relies on lazy evaluation to not evaluate \code{chunks} unless needed.}
\usage{
ragnar_store_update_document_and_chunks(store, document, chunks)
}
\description{
Workhorse that inserts a document and its chunks into the store.
Relies on lazy evaluation to not evaluate \code{chunks} unless needed.
}
