% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{ragnar_store_connect}
\alias{ragnar_store_connect}
\title{Connect to \code{RagnarStore}}
\usage{
ragnar_store_connect(location, ..., read_only = TRUE)
}
\arguments{
\item{location}{string, a filepath location.}

\item{...}{unused; must be empty.}

\item{read_only}{logical, whether the returned connection can be used to
modify the store.}
}
\value{
a \code{RagnarStore} object.
}
\description{
Connect to \code{RagnarStore}
}
