% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragnar-package.R
\docType{package}
\name{ragnar-package}
\alias{ragnar}
\alias{ragnar-package}
\title{ragnar: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://t-kalinowski.github.io/ragnar/}
}

}
\author{
\strong{Maintainer}: Tomasz Kalinowski \email{tomasz@posit.co}

}
\keyword{internal}
