% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragnar-package.R
\docType{package}
\name{ragnar-package}
\alias{ragnar}
\alias{ragnar-package}
\title{ragnar: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tomasz Kalinowski \email{tomasz@posit.co}

}
\keyword{internal}
