% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-html.R
\name{ragnar_find_links}
\alias{ragnar_find_links}
\title{Find links on a page}
\usage{
ragnar_find_links(x, depth = 0L, children_only = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{URL, HTML file path, or XML document. For Markdown, convert to HTML
using \code{\link[commonmark:commonmark]{commonmark::markdown_html()}} first.}

\item{depth}{Integer specifying how many levels deep to crawl for links. When
\code{depth > 0}, the function will follow child links (links with \code{x} as a
prefix) and collect links from those pages as well.}

\item{children_only}{Logical. If \code{TRUE}, returns only child links (those
having \code{x} as a prefix). If \code{FALSE}, returns all links found on the page.
Note that regardless of this setting, only child links are followed when
\code{depth > 0}.}

\item{verbose}{Logical, emit a \code{message()} while recursing, if \code{depth > 0}.}
}
\value{
A character vector of links on the page.
}
\description{
Find links on a page
}
\examples{
ragnar_find_links("https://r4ds.hadley.nz/base-R.html")
ragnar_find_links("https://ellmer.tidyverse.org/")
ragnar_find_links("https://ellmer.tidyverse.org/", depth = 2)
ragnar_find_links("https://ellmer.tidyverse.org/", depth = 2, children_only = FALSE)
}
