% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ragnar_retrieve}
\alias{ragnar_retrieve}
\title{Retrieve chunks from a \code{RagnarStore}}
\usage{
ragnar_retrieve(store, text, top_k = 3L)
}
\arguments{
\item{store}{A \code{RagnarStore} object.}

\item{text}{A string to find the nearest match too}

\item{top_k}{Integer, the number of nearest entries to find \emph{per method}.}
}
\value{
A dataframe of retrieved chunks. Each row corresponds to an
individual chunk in the store. It always contains a column named \code{text}
that contains the chunks.
}
\description{
\code{\link[=ragnar_retrieve]{ragnar_retrieve()}} is a thin wrapper around \code{\link[=ragnar_retrieve_vss_and_bm25]{ragnar_retrieve_vss_and_bm25()}}
using the recommended best practices.
}
\seealso{
Other ragnar_retrieve: 
\code{\link{ragnar_retrieve_bm25}()},
\code{\link{ragnar_retrieve_vss}()},
\code{\link{ragnar_retrieve_vss_and_bm25}()}
}
\concept{ragnar_retrieve}
