% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ragnar_retrieve_vss}
\alias{ragnar_retrieve_vss}
\alias{ragnar_retrieve_bm25}
\alias{ragnar_retrieve}
\title{Retreive chunks from a \code{RagnarStore}}
\usage{
ragnar_retrieve_vss(store, text, top_k = 3L)

ragnar_retrieve_bm25(store, text, top_k = 3L)

ragnar_retrieve(store, text, top_k = 3L, methods = c("vss", "bm25"))
}
\arguments{
\item{store}{A \code{RagnarStore} object.}

\item{text}{A string to find the nearest match too}

\item{top_k}{Integer, the number of nearest entries to find \emph{per method}.}

\item{methods}{A character vector of methods. \code{vss} and \code{bm25} are
implemented.}
}
\value{
A dataframe of retrieved chunks. Each row corresponds to an
individual chunk in the store, and the last column in the dataframe is
always named 'text', and contains the chunks.
}
\description{
Retreive chunks from a \code{RagnarStore}
}
