% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ragnar_retrieve_vss_and_bm25}
\alias{ragnar_retrieve_vss_and_bm25}
\title{Retrieve VSS and BM25}
\usage{
ragnar_retrieve_vss_and_bm25(store, text, top_k = 3, ...)
}
\arguments{
\item{store}{A \code{RagnarStore} object.}

\item{text}{A string to find the nearest match too}

\item{top_k}{Integer, the number of entries to retrieve using \strong{per method}.}

\item{...}{Forwarded to \code{\link[=ragnar_retrieve_vss]{ragnar_retrieve_vss()}}}
}
\value{
A dataframe of retrieved chunks. Each row corresponds to an
individual chunk in the store. It always contains a column named \code{text}
that contains the chunks.
}
\description{
Runs \code{\link[=ragnar_retrieve_vss]{ragnar_retrieve_vss()}} and \code{\link[=ragnar_retrieve_bm25]{ragnar_retrieve_bm25()}} and get the distinct
documents.
}
\note{
The results are not re-ranked after identifying the unique values.
}
\seealso{
Other ragnar_retrieve: 
\code{\link{ragnar_retrieve}()},
\code{\link{ragnar_retrieve_bm25}()},
\code{\link{ragnar_retrieve_vss}()}
}
\concept{ragnar_retrieve}
