% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{ragnar_store_create}
\alias{ragnar_store_create}
\title{Create and connect to a vector store}
\usage{
ragnar_store_create(
  location = ":memory:",
  embed = embed_ollama(),
  embedding_size = ncol(embed("foo")),
  overwrite = FALSE
)
}
\arguments{
\item{location}{filepath, or \verb{:memory:}}

\item{embed}{A function that is called with a character vector and returns a matrix of embeddings}

\item{embedding_size}{integer}

\item{overwrite}{logical, what to do if \code{location} already exists}
}
\value{
a \code{DuckDBRagnarStore} object
}
\description{
Create and connect to a vector store
}
