% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{ragnar_store_insert}
\alias{ragnar_store_insert}
\title{Insert chunks into a \code{RagnarStore}}
\usage{
ragnar_store_insert(store, chunks)
}
\arguments{
\item{store}{a \code{RagnarStore} object}

\item{chunks}{a character vector or a dataframe with a \code{text} column, and
optionally, a pre-computed \code{embedding} matrix column. If \code{embedding} is not
present, then \code{store@embed()} is used. \code{chunks} can also be a character
vector.}
}
\value{
\code{store}, invisibly.
}
\description{
Insert chunks into a \code{RagnarStore}
}
