% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{ragnar_store_inspect}
\alias{ragnar_store_inspect}
\title{Launches the Ragnar Inspector Tool}
\usage{
ragnar_store_inspect(store, ...)
}
\arguments{
\item{store}{A \code{RagnarStore} object that you want to inspect with the tool.}

\item{...}{Passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
\code{NULL} invisibly
}
\description{
Launches the Ragnar Inspector Tool
}
